chore: code to compilation file from gateway
