+++ crates/gateway/src/lib.rs
@@ -1,4 +1,5 @@
 pub mod communication;
+pub mod compilation;
 pub mod compiler_version;
 pub mod config;
 pub mod errors;
